Removes comments and moves load_css to util.py - Removes comments in query - Moves load_css function to util.py file for better management and readability
